+++ app/layout.js
@@ -1,0 +1,12 @@
+import '../styles/globals.css'
+
+export default function RootLayout({ children }) {
+  return (
+    <html lang="es">
+      <head>
+        <script src="https://cdn.tailwindcss.com"></script>
+      </head>
+      <body className="bg-gray-100">{children}</body>
+    </html>
+  )
+}
